refactor(main): centralize auto-open logic with fallback mechanism Move validation utils to shared file and improve startup project detection
