+++ app/src/main/java/com/itsaky/androidide/fragments/RecentProjectsFragment.kt
@@ -28,7 +28,6 @@ import com.itsaky.androidide.utils.flashError
 import com.itsaky.androidide.utils.viewLifecycleScope
 import com.itsaky.androidide.viewmodel.MainViewModel
 import com.itsaky.androidide.viewmodel.RecentProjectsViewModel
-import com.itsaky.androidide.preferences.internal.GeneralPreferences
 import com.itsaky.androidide.viewmodel.SortCriteria
 import com.itsaky.androidide.ui.ProjectInfoBottomSheet
 import io.sentry.Sentry
@@ -37,9 +36,12 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import org.appdevforall.codeonthego.layouteditor.ProjectFile
 import com.itsaky.androidide.utils.flashSuccess
+import com.itsaky.androidide.utils.findValidProjects
+import com.itsaky.androidide.utils.isProjectCandidateDir
+import com.itsaky.androidide.utils.isValidProjectDirectory
+import com.itsaky.androidide.utils.isValidProjectOrContainerDirectory
 import java.io.File
 
 class RecentProjectsFragment : BaseFragment() {
@@ -219,7 +221,7 @@ class RecentProjectsFragment : BaseFragment() {
 	}
 
 
-    private fun File.isProjectCandidateDir(): Boolean = isDirectory && canRead() && !name.startsWith(".") && !isHidden
+
 
     private fun bootstrapFromFixedFolderIfNeeded() {
         if (viewModel.didBootstrap) return
@@ -231,40 +233,10 @@ class RecentProjectsFragment : BaseFragment() {
                 if (validProjects.isEmpty()) return@launch
 
                 loadProjectsIntoViewModel(validProjects)
-
-                if (GeneralPreferences.autoOpenProjects) {
-                    val lastOpenedPath = GeneralPreferences.lastOpenedProject
-
-                    val projectToOpen = validProjects.find {
-                        it.absolutePath == lastOpenedPath
-                    }
-
-                    if (projectToOpen != null) {
-                        withContext(Dispatchers.Main) { openProject(projectToOpen) }
-                        return@launch
-                    }
-
-                    val lastCreated = validProjects.maxByOrNull { it.lastModified() }
-
-                    if (lastCreated != null) {
-                        withContext(Dispatchers.Main) { openProject(lastCreated) }
-                    }
-                }
             } catch (e: Throwable) {
                 Sentry.captureException(e)
             }
         }
-    }
-
-    private fun findValidProjects(projectsRoot: File): List<File> {
-        if (!projectsRoot.isProjectCandidateDir()) return emptyList()
-
-        val subdirs = projectsRoot.listFiles()
-            ?.filter { it.isProjectCandidateDir() }
-            .orEmpty()
-        if (subdirs.isEmpty()) return emptyList()
-
-        return subdirs.filter { dir -> isValidProjectDirectory(dir) }
     }
 
     private suspend fun loadProjectsIntoViewModel(projects: List<File>) {
@@ -394,43 +366,6 @@ class RecentProjectsFragment : BaseFragment() {
                 }
             }
         }
-    }
-
-	fun isValidProjectDirectory(selectedDir: File): Boolean {
-        if (isPluginProject(selectedDir)) {
-            return true
-        }
-
-        val appFolder = File(selectedDir, "app")
-        val buildGradleFile = File(appFolder, "build.gradle")
-        val buildGradleKtsFile = File(appFolder, "build.gradle.kts")
-        return appFolder.exists() && appFolder.isDirectory &&
-                (buildGradleFile.exists() || buildGradleKtsFile.exists())
-    }
-
-    private fun isPluginProject(dir: File): Boolean {
-        val pluginApiJar = File(dir, "libs/plugin-api.jar")
-        val buildGradle = File(dir, "build.gradle.kts")
-        return pluginApiJar.exists() && buildGradle.exists()
-    }
-
-    /**
-     * Determines if the selected directory is either:
-     *  1. A valid Android project itself, OR
-     *  2. A container that includes one or more valid Android projects.
-     */
-    fun isValidProjectOrContainerDirectory(selectedDir: File): Boolean {
-        if (!selectedDir.isProjectCandidateDir()) {
-            return false
-        }
-
-        if (isValidProjectDirectory(selectedDir)) {
-            return true
-        }
-
-        // Check if it contains valid Android projects as subdirectories
-        val subDirs = selectedDir.listFiles()?.filter { it.isProjectCandidateDir() } ?: return false
-        return subDirs.any { sub -> isValidProjectDirectory(sub) }
     }
 
 		private fun openProjectInfo(project: ProjectFile) {
